Rewrite rest of classes, increased engine precision
