--- examples/07_chain.html
@@ -1,53 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Chain</title>
-    <script src="../build/end3in.js"></script>
-</head>
-<body></body>
-<script>
-    let Rectangle = end3in.Rectangle,
-        Circle = end3in.Circle,
-        Joint = end3in.Joint,
-        Engine = end3in.Engine,
-        Graphics = end3in.Graphics,
-        Vec2 = end3in.Vec2;
-
-    engine = new Engine(true, 'constraint');
-    engine.scale = 50;
-
-
-
-    let bodies = [new Rectangle(3, 0.25, 0.25, 0.125)];
-    let joints = [];
-    let lastBody = bodies[0];
-    for (let i = 0; i < 10; i++) {
-        let body = new Circle(0, 0, 0.25, 1);
-        bodies.push(body);
-        let joint = new Joint(lastBody, body, new Vec2(i, i));
-        joints.push(joint);
-        lastBody = body;
-    }
-    engine.addAllBodies(bodies);
-    engine.addAllJoints(joints);
-    engine.g = 0;
-    engine.start();
-
-    const graphics = new Graphics(1000, 1000, engine)
-    graphics.start()
-
-    function setup() {
-        for (let i = 1; i < bodies.length; i++) {
-            bodies[i].pos = new Vec2(i * 0.5, i * 0.5);
-            bodies[i].v = new Vec2(0, 0);
-        }
-    }
-
-    setup();
-
-    document.onmousedown = function () {
-        setup()
-    }
-</script>
-</html>
