+++ src/main.ts
@@ -3,5 +3,4 @@ export { Collision } from './Collision'
 export { Polygon, Circle, RegularPolygon, Rectangle } from './Body'
 export { Engine } from './Engine'
 export { Solver } from './Solver'
-export { Joint } from './Joint'
 export { Graphics } from './Graphics'
